requirements.txt
Django==1.8
psycopg2==2.5.3
django-pagination-bootstrap==1.0.7
django-pagination==1.0.7
pytils
django-uuslug
webstack-django-sorting

ipdb==0.8
django_nose
coverage
factory-boy==2.4.1
django-debug-toolbar

gunicorn